+++ source/components/Bubble.js
@@ -12,9 +12,16 @@ import { addDocumentListeners, removeDocumentListeners } from '../utils/events';
 import { IDENTIFIERS } from '.';
 import type { ThemeContextProp } from './HOC/withTheme';
 
-type Props = {
+export type BubblePosition = {
+  width: number,
+  positionX: number,
+  positionY: number,
+};
+
+export type BubbleProps = {
   className?: string,
   context: ThemeContextProp,
+  isCentered: boolean,
   isHidden: boolean,
   isFloating: boolean,
   isOpeningUpward: boolean,
@@ -30,10 +37,10 @@ type Props = {
 
 type State = {
   composedTheme: Object,
-  position: ?Object
+  position: ?BubblePosition
 };
 
-class BubbleBase extends Component<Props, State> {
+class BubbleBase extends Component<BubbleProps, State> {
   // declare ref types
   rootElement: ?Element<any>;
 
@@ -41,6 +48,7 @@ class BubbleBase extends Component<Props, State> {
   static displayName = 'Bubble';
   static defaultProps = {
     context: createEmptyContext(),
+    isCentered: false,
     isHidden: false,
     isFloating: false,
     isOpeningUpward: false,
@@ -52,7 +60,7 @@ class BubbleBase extends Component<Props, State> {
     themeOverrides: {}
   };
 
-  constructor(props: Props) {
+  constructor(props: BubbleProps) {
     super(props);
 
     // define ref
@@ -78,7 +86,7 @@ class BubbleBase extends Component<Props, State> {
     }, 0);
   }
 
-  componentWillReceiveProps(nextProps: Props) {
+  componentWillReceiveProps(nextProps: BubbleProps) {
     didThemePropsChange(this.props, nextProps, this.setState.bind(this));
   }