+++ source/skins/simple/TooltipSkin.js
@@ -1,12 +1,12 @@
 // @flow
 import React from 'react';
-import type { Node, Element } from 'react';
 
 // external libraries
 import classnames from 'classnames';
 
 // components
 import { Bubble } from '../../components/Bubble';
+import type { TooltipProps } from '../../components/Tooltip';
 
 // skins
 import { BubbleSkin } from './BubbleSkin';
@@ -14,17 +14,9 @@ import { BubbleSkin } from './BubbleSkin';
 // internal utility functions
 import { pickDOMProps } from '../../utils/props';
 
-type Props = {
-  children?: ?Node,
-  className?: string,
-  isAligningRight?: boolean,
-  isBounded?: boolean,
-  isOpeningUpward: boolean,
-  isTransparent: boolean,
-  arrowRelativeToTip: boolean,
+type Props = TooltipProps & {
   theme: Object,
   themeId: string,
-  tip: string | Element<any>
 };
 
 export const TooltipSkin = (props: Props) => {
@@ -32,7 +24,12 @@ export const TooltipSkin = (props: Props) => {
   return (
     <span
       {...pickDOMProps(props)}
-      className={classnames([props.className, theme[themeId].root])}
+      className={classnames([
+        props.className,
+        theme[themeId].root,
+        props.isVisible ? theme[themeId].isVisible : null,
+        props.isCentered ? theme[themeId].isCentered : null,
+      ])}
     >
       <Bubble
         className={classnames([
@@ -43,6 +40,7 @@ export const TooltipSkin = (props: Props) => {
           props.isBounded ? null : theme[themeId].nowrap
         ])}
         theme={theme}
+        isCentered={props.isCentered}
         isOpeningUpward={props.isOpeningUpward}
         skin={BubbleSkin}
         isTransparent={props.isTransparent}